Add fallback implementation for `GetProcessId`
Since that API is only available starting from XP SP1/Server 2003, we'll
have to get a bit creative to emulate the same behavior on older
systems:

We'll have to query whether the process has exited by
`WaitForSingleObject` with a timeout of 0 (= non-blocking). If
`WAIT_TIMEOUT` is returned, the process has not exited, so our saved
process id is still valid (MSDN: "The value is valid from the time the
process is created until the time the process is terminated."). If the
process has exited, the wait will succeed and return `WAIT_OBJECT_0`, in
which case the id is no longer valid, and we return 0, just like
`GetProcessId`.