Stub `SetHandleInformation`/`Socket::set_no_inherit`
SetHandleInformation is exported by kernel32 on Win9X/ME, but is just
stubbed, returning `ERROR_CALL_NOT_IMPLEMENTED`. Sockets are
non-inheritable on these systems anyways, so we "fail successfully".
https://www.betaarchive.com/wiki/index.php/Microsoft_KB_Archive/150523#MORE_INFORMATION

SetHandleInformation is also unavailable on WinNT before 3.51. This is
fine, however, because MS did not supply WinSock 2 for Windows NT before
4.0, so this function is not called.