Add fallback implementation for `CopyFileExW`
If `CopyFileExW` is not available, we have to
- copy the file with the non-Ex API
- then open it with `dwDesiredAccess = 0` (query attributes only)
- then use `GetFileSize` to retrieve the size

For some reason, the unicows layer implements `CopyFileExW` similarly to
other functions (convert to ANSI, call ...A API). However, 9x/ME don't
support `CopyFileExA` either! This means we have to check both for the
API to exist *and* that we're running on NT.