Add fallback implementation in stdio redirection for when named pipes are not available
Since Windows 9X/ME does not support creating named pipes (only
connecting to remote pipes created on NT), we'll have to make do with
anonymous pipes, without overlapped I/O. In particular, this means that
we'll have to spawn another thread in the case where both stdout and
stderr being piped.

We also use the fallback implementation on NT before 4.0, as the `Drop`
impl of `AsyncPipe` needs to be able to cancel I/O via `CancelIo`.